clean-up warnings
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@1186 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ addon/ms-logon/MSLogonACL/vncImportACL.h
@@ -80,7 +80,7 @@ struct vncScanInput{
 	void RemoveComment(TCHAR *line);
 	void RemoveNewline(TCHAR *line);
 	int GetQuoteLength(const TCHAR *line);
-	int GetWordLength(const TCHAR *line);
+    size_t GetWordLength(const TCHAR *line);
 	bool GetLine(TCHAR *line);
 	bool Tokenize(const TCHAR *line, TCHAR **tokens);
 	int AddToken(TCHAR **token, int tokenCount, const TCHAR **line, int len);
